+++ src/Flare.Theme.MaterialDesign3.Tokens/MaterialDesign3Tokens.cs
@@ -967,6 +967,8 @@ public class MaterialDesign3Tokens
         StateLayerSize = "40px",
         StateHoverOpacity = "0.08",
         StatePressedOpacity = "0.10",
+        FocusOutline = "none",
+        FocusOutlineOffset = "0px",
         StopColor = "var(--flare-color-on-secondary-container)",
         StopColorSelected = "var(--flare-color-on-primary)",
         StopSize = "4px",